+++ src/logger/job-logger.mjs
@@ -10,14 +10,28 @@ export class JobLogger extends AbstractLogger {
     }
 
     async info(message) {
-        await this.#job.log(`Info: ${message}`);
+        await this.#job.log(`[info]: ${message}`);
     }
 
     async warning(message) {
-        await this.#job.log(`Warning: ${message}`);
+        await this.#job.log(`[warning]: ${this.#formatErrorMessage(message)}`);
     }
 
     async error(message) {
-        await this.#job.log(`Error: ${message}`);
+        await this.#job.log(`[error]: ${this.#formatErrorMessage(message)}`);
+    }
+
+    #formatErrorMessage(message) {
+        if (message instanceof Error) {
+            let errorMessage = message.toString();
+
+            if (!errorMessage.endsWith('.')) {
+                errorMessage += '.';
+            }
+
+            return `${errorMessage} Stack:\n${message.stack}`;
+        }
+
+        return message;
     }
 }